Switch to larger number types

+++ include/cpp/jank/runtime/detail/type.hpp
@@ -13,8 +13,8 @@
 namespace jank::runtime::detail
 {
   using memory_policy = immer::memory_policy<immer::free_list_heap_policy<immer::cpp_heap>, immer::refcount_policy, immer::default_lock_policy>;
-  using integer_type = int64_t;
-  using real_type = double;
+  using integer_type = long long;
+  using real_type = long double;
   using boolean_type = bool;
   using string_type = string_type_impl<memory_policy>;
